Section 20: File Upload & Download

+++ controllers/admin.js
@@ -2,6 +2,8 @@ const { validationResult } = require('express-validator/check')
 
 const Product = require('../models/product')
 
+const fileHelper = require('../util/file')
+
 exports.getAddProduct = (req, res, next) => {
 
     let message = req.flash('error')
@@ -121,16 +123,11 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId
-    const updatedTitle = req.body.title
-    const updatedPrice = req.body.price
     const image = req.file
-    const updatedDesc = req.body.description
-
     const title = req.body.title;
     const imageUrl = req.body.imageUrl;
     const price = req.body.price;
     const description =req.body.description;
-
     const errors = validationResult(req)
 
     if(!errors.isEmpty()) {
@@ -154,13 +151,14 @@ exports.postEditProduct = (req, res, next) => {
             if (product.userId.toString() !== req.user._id.toString()) {
                 return res.redirect('/')
             }
-            product.title = updatedTitle
-            product.price = updatedPrice
-            product.description = updatedDesc
+            product.title = title
+            product.price = price
+            product.description = description
             if (image) {
+                fileHelper.deleteFile(product.imageUrl)
                 product.imageUrl = image.path
             }
-            product.imageUrl = updatedImageUrl
+            product.imageUrl = imageUrl
             return product.save().then(result => { 
                 console.log('UPDATED PRODUCT!');
                 res.redirect('/admin/products');
@@ -175,8 +173,6 @@ exports.postEditProduct = (req, res, next) => {
       
 exports.getProducts = (req, res, next) => {
     Product.find({ userId: req.user._id })
-    // .select('title price -_id')
-    // .populate('userId', 'name')
         .then(products => {
             console.log(products)
             res.render('admin/products', {
@@ -191,16 +187,23 @@ exports.getProducts = (req, res, next) => {
         return next(error)
     });
 };
+
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId
-    Product.deleteOne({_id: prodId, userId: req.user._id })
-        .then(() => {
-            console.log('Destroyed Product')
-            res.redirect('/admin/products')
-        })
-        .catch(err => {
-            const error = new Error(err)
-            error.httpStatusCode = 500
-            return next(error)
-        })
-    }    
+    Product.findById(prodId).then(product => {
+        if (!product) {
+            return next(new Error('Product not found.'))
+        }
+        fileHelper.deleteFile(product.imageUrl)
+        return Product.deleteOne({_id: prodId, userId: req.user._id })
+    })
+    .then(() => {
+        console.log('Destroyed Product')
+        res.redirect('/admin/products')
+    })
+    .catch(err => {
+        const error = new Error(err)
+        error.httpStatusCode = 500
+        return next(error)
+    })
+}
